US106353 - extract sorting
